SKALE-3030 Rename config_builder module
