+++ core/schains/config/accounts.py
@@ -22,7 +22,7 @@ import logging
 from web3 import Web3
 from skale.wallets.web3_wallet import public_key_to_address
 
-from core.schains.config_builder.helper import fix_address, _string_to_storage, get_context_contract
+from core.schains.config.helper import fix_address, _string_to_storage, get_context_contract
 from core.schains.filestorage import compose_filestorage_info
 from core.schains.helper import read_ima_data
 from core.schains.volume import get_resource_allocation_info, get_allocation_part_name
