+++ tests/schains/config/generator.py
@@ -1,6 +1,6 @@
 import json
-from core.schains.config_builder.generator import (generate_schain_config,
-                                                   generate_schain_config_with_skale)
+from core.schains.config.generator import (generate_schain_config,
+                                           generate_schain_config_with_skale)
 
 
 SCHAIN_NODES_WITH_SCHAINS = [{'name': 'test_node_2', 'ip': b'L\xea\xbb\xc9', 'publicIP': b'\xb6w\x01\xfe', 'port': 26155, 'start_block': 236, 'last_reward_date': 1599240149, 'finish_time': 0, 'status': 0, 'validator_id': 1, 'publicKey': '0xf925c203a30ec6cad5a263db3efab7ed4c1fd74c8688167e10a5a22e15ab5018d8553df0ac54ea105a3d21845e5660bc3d4e7c82e7af1daa3baad393b1521467', 'id': 1, 'schains': [{'name': 'test', 'owner': '0x5112cE768917E907191557D7E9521c2590Cdd3A0', 'indexInOwnerList': 0, 'partOfNode': 0, 'lifetime': 3600, 'startDate': 1599240155, 'startBlock': 240, 'deposit': 1000000000000000000, 'index': 0, 'chainId': '0x9ca5dee9297f2', 'active': True}], 'bls_public_key': ((0, 0), (1, 0))}, {'name': 'test_node', 'ip': b'\xbc\xbed\xb3', 'publicIP': b'\xcd\xa1\xf4g', 'port': 43118, 'start_block': 232, 'last_reward_date': 1599240145, 'finish_time': 0, 'status': 0, 'validator_id': 1, 'publicKey': '0xf925c203a30ec6cad5a263db3efab7ed4c1fd74c8688167e10a5a22e15ab5018d8553df0ac54ea105a3d21845e5660bc3d4e7c82e7af1daa3baad393b1521467', 'id': 0, 'schains': [{'name': 'test', 'owner': '0x5112cE768917E907191557D7E9521c2590Cdd3A0', 'indexInOwnerList': 0, 'partOfNode': 0, 'lifetime': 3600, 'startDate': 1599240155, 'startBlock': 240, 'deposit': 1000000000000000000, 'index': 0, 'chainId': '0x9ca5dee9297f2', 'active': True}], 'bls_public_key': ((0, 0), (1, 0))}]  # noqa
